refactor(#166): hash.ts 통합
